+++ packages/stats-validator-plugin-webpack/src/rules/diff-entry-download-time-limits/index.ts
@@ -30,8 +30,7 @@ export type NormalizedLimits = {
   maxAsyncDownloadTimeDiff: Limit | null;
 };
 
-export type RuleExcludeItem = ExcludeItem<'compilation' | 'entry'>;
-export type SerializedRuleExcludeItem = SerializedExcludeItem<'compilation' | 'entry'>;
+export type RuleExcludeItem = ExcludeItem<'compilation' | 'entry' | 'asset'>;
 
 export type Params = {
   useCompressedSize?: boolean;
@@ -156,7 +155,10 @@ const diffEntryDownloadTimeLimits: WebpackRule<Params> = (
   $params: #.params;
   $useCompressedSize: [$params.useCompressedSize, true].useNotNullish();
   $network: [$params.network, 'Slow'].useNotNullish();
-  $getSizeByChunks: => files.(getAssetSize($$, $useCompressedSize!=false)).reduce(=> size + $$, 0);
+  $getSizeByChunks: => files.exclude({
+    exclude: $params.exclude.[type='asset'].name,
+    get: <name>,
+  }).(getAssetSize($$, $useCompressedSize!=false)).reduce(=> size + $$, 0);
   
   $input.compilations
   .exclude({