+++ packages/webpack-ui/src/init/init-menu.ts
@@ -1,10 +1,9 @@
-import { ViewConfigData } from '@statoscope/types';
 import networkTypeList from '@statoscope/helpers/dist/network-type-list';
-// @ts-ignore
-import settingsStyles from '../settings-styles.css';
-import settings, {
+import {
   SETTING_ASSETS_INJECT_TYPE,
   SETTING_ASSETS_INJECT_TYPE_DEFAULT,
+  SETTING_EXCLUDE_RESOURCES_FROM_SIZE_CALC,
+  SETTING_EXCLUDE_RESOURCES_FROM_SIZE_CALC_DEFAULT,
   SETTING_HIDE_CHILD_COMPILATIONS,
   SETTING_HIDE_CHILD_COMPILATIONS_DEFAULT,
   SETTING_HIDE_NODE_MODULES,
@@ -17,6 +16,12 @@ import settings, {
   SETTING_SHOW_COMPRESSED_DEFAULT,
 } from '../settings';
 import { StatoscopeWidget } from '../../types';
+import {
+  makeBooleanSetting,
+  makeSelectSetting,
+  makeStringSetting,
+  makeToggleSetting,
+} from './settings';
 
 export default (discovery: StatoscopeWidget): void => {
   hideUseless(discovery);
@@ -131,73 +136,6 @@ function addReportsList(discovery: StatoscopeWidget): void {
   });
 }
 
-export type SettingOptions<TValue> = {
-  title: string;
-  key: string;
-  hint?: string;
-  defaultValue: TValue;
-};
-
-function makeBooleanSetting(
-  discovery: StatoscopeWidget,
-  { title, key, hint, defaultValue }: SettingOptions<boolean>
-): ViewConfigData {
-  return {
-    view: 'block',
-    className: [settingsStyles.item, settingsStyles.toggle],
-    name: key,
-    postRender: (
-      el: HTMLElement,
-      opts: unknown,
-      data: unknown,
-      { hide }: { hide(): void }
-    ): void => {
-      const settingValue = settings.get(key, defaultValue);
-
-      render(settingValue.get());
-
-      settingValue.eventChange.on((sender, { value }) => render(value || false));
-
-      function render(value: boolean): void {
-        el.innerHTML = '';
-        discovery.view.render(
-          el,
-          // @ts-ignore
-          {
-            view: 'toggle-group',
-            beforeToggles: [
-              `text:"${title}"`,
-              (
-                element: HTMLDivElement /*,
-                props: unknown,
-                data: Array<{ value: boolean; text: string }>,
-                context: { widget: StatoscopeWidget }*/
-              ): void => {
-                if (hint) {
-                  element.style.display = 'inline-block';
-                  element.textContent = '❓';
-                  element.title = hint;
-                }
-              },
-            ],
-            onChange: (value: boolean) => {
-              settingValue.set(value);
-              hide();
-            },
-            value,
-            data: [
-              { value: false, text: 'No' },
-              { value: true, text: 'Yes' },
-            ],
-          },
-          null,
-          { widget: discovery }
-        );
-      }
-    },
-  };
-}
-
 function addSettings(discovery: StatoscopeWidget): void {
   discovery.nav.menu.append(
     makeBooleanSetting(discovery, {
@@ -224,157 +162,60 @@ function addSettings(discovery: StatoscopeWidget): void {
     })
   );
 
-  discovery.nav.menu.append({
-    view: 'block',
-    className: [settingsStyles.item, settingsStyles.toggle],
-    name: 'items-limit',
-    postRender: (
-      el: HTMLElement,
-      opts: unknown,
-      data: unknown,
-      { hide }: { hide(): void }
-    ): void => {
-      const limit = settings.get<string>(
-        SETTING_LIST_ITEMS_LIMIT,
-        SETTING_LIST_ITEMS_LIMIT_DEFAULT
-      );
+  discovery.nav.menu.append(
+    makeToggleSetting(
+      discovery,
+      {
+        title: 'List items limit',
+        key: SETTING_LIST_ITEMS_LIMIT,
+        defaultValue: SETTING_LIST_ITEMS_LIMIT_DEFAULT,
+      },
+      [
+        { value: '10', text: '10' },
+        { value: '20', text: '20' },
+        { value: '50', text: '50' },
+        { value: '100', text: '100' },
+      ]
+    )
+  );
 
-      render(limit.get());
+  discovery.nav.menu.append(
+    makeSelectSetting(
+      discovery,
+      {
+        title: 'Network type',
+        key: SETTING_NETWORK_SPEED,
+        defaultValue: SETTING_NETWORK_SPEED_DEFAULT,
+      },
+      networkTypeList.map((item) => item.name),
+      'settingNetworkType()',
+      'getNetworkTypeInfo().getNetworkTypeName()'
+    )
+  );
 
-      limit.eventChange.on((sender, { value }) => render(value));
+  discovery.nav.menu.append(
+    makeToggleSetting(
+      discovery,
+      {
+        title: 'Assets inject type',
+        hint: 'sync: download time = sum(downloadTime(assets))\nasync: dowload time = max(downloadTime(assets))',
+        key: SETTING_ASSETS_INJECT_TYPE,
+        defaultValue: SETTING_ASSETS_INJECT_TYPE_DEFAULT,
+      },
+      ['sync', 'async']
+    )
+  );
 
-      function render(value: string): void {
-        el.innerHTML = '';
-        discovery.view.render(
-          el,
-          // @ts-ignore
-          {
-            view: 'toggle-group',
-            beforeToggles: 'text:"List items limit"',
-            onChange: (value: string) => {
-              limit.set(value);
-              hide();
-            },
-            value,
-            data: [
-              { value: '10', text: 10 },
-              { value: '20', text: 20 },
-              { value: '50', text: 50 },
-              { value: '100', text: 100 },
-            ],
-          },
-          null,
-          { widget: discovery }
-        );
-      }
-    },
-  });
-
-  discovery.nav.menu.append({
-    view: 'block',
-    className: [settingsStyles.item, settingsStyles.select],
-    name: 'network-type',
-    postRender: (
-      el: HTMLElement,
-      opts: unknown,
-      data: unknown,
-      { hide }: { hide(): void }
-    ): void => {
-      const limit = settings.get<typeof networkTypeList[number]['name']>(
-        SETTING_NETWORK_SPEED,
-        SETTING_NETWORK_SPEED_DEFAULT
-      );
-
-      render();
-
-      limit.eventChange.on(() => render());
-
-      function render(): void {
-        el.innerHTML = '';
-        discovery.view.render(
-          el,
-          [
-            {
-              view: 'block',
-              className: settingsStyles.caption,
-              content: `text:"Network type"`,
-            },
-            {
-              view: 'select',
-              placeholder: 'choose a type',
-              value: 'settingNetworkType()',
-              text: `getNetworkTypeInfo().getNetworkTypeName()`,
-              data: networkTypeList.map((item) => item.name),
-              onChange: (value: typeof networkTypeList[number]['name']): void => {
-                limit.set(value);
-                hide();
-              },
-            },
-          ],
-          null,
-          { widget: discovery }
-        );
-      }
-    },
-  });
-
-  discovery.nav.menu.append({
-    view: 'block',
-    className: [settingsStyles.item, settingsStyles.select],
-    name: 'client-download-type',
-    postRender: (
-      el: HTMLElement,
-      opts: unknown,
-      data: unknown,
-      { hide }: { hide(): void }
-    ): void => {
-      const limit = settings.get(
-        SETTING_ASSETS_INJECT_TYPE,
-        SETTING_ASSETS_INJECT_TYPE_DEFAULT
-      );
-
-      render();
-
-      limit.eventChange.on(() => render());
-
-      function render(): void {
-        el.innerHTML = '';
-        discovery.view.render(
-          el,
-          [
-            {
-              view: 'block',
-              className: settingsStyles.caption,
-              content: [
-                `text:"Assets inject type"`,
-                (
-                  element: HTMLDivElement /*,
-                props: unknown,
-                data: Array<{ value: boolean; text: string }>,
-                context: { widget: StatoscopeWidget }*/
-                ): void => {
-                  element.style.display = 'inline-block';
-                  element.textContent = '❓';
-                  element.title =
-                    'sync: dowload time = sum(downloadTime(assets))\nasync: dowload time = max(downloadTime(assets))';
-                },
-              ],
-            },
-            {
-              view: 'select',
-              placeholder: 'choose a type',
-              value: 'settingAssetsInjectType()',
-              data: ['sync', 'async'],
-              onChange: (value: string): void => {
-                limit.set(value);
-                hide();
-              },
-            },
-          ],
-          null,
-          { widget: discovery }
-        );
-      }
-    },
-  });
+  discovery.nav.menu.append(
+    makeStringSetting(
+      discovery,
+      {
+        title: 'Ignore from size calculation',
+        hint: 'Ignore matched resources from size calculation',
+        key: SETTING_EXCLUDE_RESOURCES_FROM_SIZE_CALC,
+        defaultValue: SETTING_EXCLUDE_RESOURCES_FROM_SIZE_CALC_DEFAULT,
+      },
+      'regexp'
+    )
+  );
 }